+++ module/gurps.js
@@ -10,6 +10,7 @@ import {
   GurpsActorEditorSheet,
   GurpsActorSimplifiedSheet,
   GurpsActorNpcSheet,
+  GurpsInventorySheet
 } from './actor-sheet.js'
 import { ModifierBucket } from './modifier-bucket/bucket-app.js'
 import { ChangeLogWindow } from '../lib/change-log.js'
@@ -20,6 +21,7 @@ import { doRoll } from '../module/dierolls/dieroll.js'
 import { ResourceTrackerManager } from './actor/resource-tracker-manager.js'
 import { DamageTables, initializeDamageTables } from '../module/damage/damage-tables.js'
 import RegisterChatProcessors from '../module/chat/chat-processors.js'
+import { Migration } from '../lib/migration.js'
 
 export const GURPS = {}
 window.GURPS = GURPS // Make GURPS global!
@@ -338,7 +340,7 @@ CONFIG.statusEffects = [
   },
   {
     icon: 'systems/gurps/icons/statuses/dd-condition-unconscious.png',
-    id: 'disbled',
+    id: 'disabled',
     label: 'GURPS.STATUSDisable',
   },
   {
@@ -627,12 +629,15 @@ async function performAction(action, actor, event, targets) {
     let chat = action.orig
     if (!!event?.shiftKey || game.keyboard.isCtrl(event))
       chat = `/setEventFlags ${!!event?.shiftKey} ${game.keyboard.isCtrl(event)}\n${chat}`
-    ui.chat.processMessage(chat).catch(err => {
+      
+    return await GURPS.ChatProcessors.startProcessingLines(chat, event?.chatmsgData, event)
+/*    ui.chat.processMessage(chat).catch(err => {
       ui.notifications.error(err)
       console.error(err)
       return false
     })
     return true
+    */
   }
 
   if (action.type === 'controlroll') {
@@ -701,7 +706,7 @@ async function performAction(action, actor, event, targets) {
     while (!!tempAction) {
       if (!!tempAction.truetext && !besttrue) besttrue = tempAction
       if (tempAction.type == 'attribute') {
-        th = this.i18n(tempAction.path)
+        th = this._mapAttributePath(tempAction.path)
         let t = parseInt(tempAction.target) // is it pre-targeted (ST12)
         if (!t && !!actor) {
           if (!!tempAction.melee) {
@@ -868,6 +873,7 @@ function findSkillSpell(actor, sname, isSkillOnly = false, isSpellOnly = false) 
   if (!actor) return t
   if (!!actor.data?.data?.additionalresources) actor = actor.data
   sname = makeRegexPatternFrom(sname, false)
+  sname = new RegExp(sname, "i");
   let best = 0
   if (!isSpellOnly)
     recurselist(actor.data.skills, s => {
@@ -893,6 +899,7 @@ function findAdDisad(actor, sname) {
   if (!actor) return t
   if (!!actor.data?.data?.additionalresources) actor = actor.data
   sname = makeRegexPatternFrom(sname, false)
+  sname = new RegExp(sname, "i");
   recurselist(actor.data.ads, s => {
     if (s.name.match(sname)) {
       t = s
@@ -907,6 +914,7 @@ function findAttack(actor, sname, isMelee = true, isRanged = true) {
   if (!actor) return t
   if (!!actor.data?.data?.additionalresources) actor = actor.data
   sname = makeRegexPatternFrom(sname, false)
+  sname = new RegExp(sname, "i");
   if (isMelee)
     t = actor.data.melee?.findInProperties(a => (a.name + (!!a.mode ? ' (' + a.mode + ')' : '')).match(sname))
   if (isRanged && !t)
@@ -941,7 +949,7 @@ async function handleRoll(event, actor, targets) {
     return
   } else if ('path' in element.dataset) {
     prefix = 'Roll vs '
-    thing = this.i18n(element.dataset.path)
+    thing = this._mapAttributePath(element.dataset.path)
     formula = '3d6'
     target = parseInt(element.innerText)
   } else if ('name' in element.dataset || 'otf' in element.dataset) {
@@ -1075,16 +1083,15 @@ GURPS.handlePdf = handlePdf
 
 // Return the i18n string for this data path (note en.json must match up to the data paths).
 // special case, drop ".value" from end of path (and append "NAME"), usually used for attributes
-// function _i18n(path, suffix) {
-//   let i = path.indexOf('.value')
-//   if (i >= 0) {
-//     path = path.substr(0, i) + 'NAME' // used for the attributes
-//   }
-
-//   path = path.replace(/\./g, '') // remove periods
-//   return game.i18n.localize('GURPS.' + path)
-// }
-// GURPS._i18n = _i18n
+ function _mapAttributePath(path, suffix) {
+   let i = path.indexOf('.value')
+   if (i >= 0) {
+     path = path.substr(0, i) + 'NAME' // used for the attributes
+   }
+   path = path.replace(/\./g, '') // remove periods
+   return game.i18n.localize('GURPS.' + path)
+ }
+ GURPS._mapAttributePath = _mapAttributePath
 
 // Given a string path "x.y.z", use it to resolve down an object heiracrhy
 function resolve(path, obj = self, separator = '.') {
@@ -1163,6 +1170,8 @@ async function removeKey(actor, path) {
   let objkey = objpath.substr(i + 1)
   let object = GURPS.decode(actor.data, objpath)
   let t = parentpath + '.-=' + objkey
+  let oldRender = actor.ignoreRender
+  actor.ignoreRender = true
   await actor.update({ [t]: null }) // Delete the whole object
   delete object[key]
   i = parseInt(key)
@@ -1181,6 +1190,7 @@ async function removeKey(actor, path) {
       a[v] = object[v]
       return a
     }, {}) // Enforced key order
+  actor.ignoreRender = oldRender
   await actor.update({ [objpath]: sorted })
 }
 GURPS.removeKey = removeKey
@@ -1519,6 +1529,10 @@ Hooks.once('init', async function () {
     label: 'NPC/mini',
     makeDefault: false,
   })
+  Actors.registerSheet('gurps', GurpsInventorySheet, {
+      label: 'Inventory Only',
+      makeDefault: false,
+  })
 
   Items.unregisterSheet('core', ItemSheet)
   Items.registerSheet('gurps', GurpsItemSheet, { makeDefault: true })
@@ -1542,13 +1556,23 @@ Hooks.once('ready', async function () {
     template: 'systems/gurps/templates/threed6.html',
     classes: [],
   }).render(true)
+  
+  GURPS.currentVersion = SemanticVersion.fromString(game.system.data.version)
+  // Test for migration
+  const mv = game.settings.get(settings.SYSTEM_NAME, settings.SETTING_MIGRATION_VERSION) || '0.0.1'
+  console.log("Current Version: " + GURPS.currentVersion + ", Migration version: " + mv)
+  const migrationVersion = SemanticVersion.fromString(mv)
+  if (migrationVersion.isLowerThan(GURPS.currentVersion)) { // check which migrations are needed
+    if (migrationVersion.isLowerThan(settings.VERSION_096)) await Migration.migrateTo096()
+    if (migrationVersion.isLowerThan(settings.VERSION_097)) await Migration.migrateTo097()
+    game.settings.set(settings.SYSTEM_NAME, settings.SETTING_MIGRATION_VERSION, game.system.data.version)
+  }
 
   // Show changelog
   const v = game.settings.get(settings.SYSTEM_NAME, settings.SETTING_CHANGELOG_VERSION) || '0.0.1'
   const changelogVersion = SemanticVersion.fromString(v)
-  const curVersion = SemanticVersion.fromString(game.system.data.version)
 
-  if (curVersion.isHigherThan(changelogVersion)) {
+  if (GURPS.currentVersion.isHigherThan(changelogVersion)) {
     if ($(ui.chat.element).find('#GURPS-LEGAL').length == 0)
       // If it isn't already in the chat log somewhere
       ChatMessage.create({
@@ -1569,7 +1593,7 @@ Hooks.once('ready', async function () {
     if (game.settings.get(settings.SYSTEM_NAME, settings.SETTING_SHOW_CHANGELOG)) {
       const app = new ChangeLogWindow(changelogVersion)
       app.render(true)
-      game.settings.set(settings.SYSTEM_NAME, settings.SETTING_CHANGELOG_VERSION, curVersion.toString())
+      game.settings.set(settings.SYSTEM_NAME, settings.SETTING_CHANGELOG_VERSION, GURPS.currentVersion.toString())
     }
   }
 
@@ -1687,9 +1711,16 @@ Hooks.once('ready', async function () {
       let srcActor = game.actors.get(resp.srcactorid)
       Dialog.confirm({
         title: `Gift for ${destactor.name}!`,
-        content: `<p>${srcActor.name} wants to give you ${resp.itemData.name},</p><br>Ok?`,
+        content: `<p>${srcActor.name} wants to give you ${resp.itemData.name} (${resp.count}),</p><br>Ok?`,
         yes: () => {
-          destactor.addNewItemData(resp.itemData)
+        
+          let destKey = destactor._findEqtkeyForGlobalItem(resp.itemData.data.globalid)
+          if (!!destKey) {    // already have some
+            let destEqt = getProperty(destactor.data, destKey)
+            destactor.updateEqtCount(destKey, destEqt.count + resp.count)
+          } else {
+            destactor.addNewItemData(resp.itemData)
+          }
           game.socket.emit('system.gurps', {
             type: 'dragEquipment2',
             srckey: resp.srckey,
@@ -1697,6 +1728,7 @@ Hooks.once('ready', async function () {
             srcactorid: resp.srcactorid,
             destactorid: resp.destactorid,
             itemname: resp.itemData.name,
+            count: resp.count
           })
         },
         no: () => {
@@ -1712,7 +1744,12 @@ Hooks.once('ready', async function () {
     if (resp.type == 'dragEquipment2') {
       if (resp.srcuserid != game.user.id) return
       let srcActor = game.actors.get(resp.srcactorid)
-      srcActor.deleteEquipment(resp.srckey)
+      let eqt = getProperty(srcActor.data, resp.srckey)
+      if (resp.count >= eqt.count) {
+        srcActor.deleteEquipment(resp.srckey)
+      } else { 
+        srcActor.updateEqtCount(resp.srckey, eqt.count - resp.count)
+      }
       let destActor = game.actors.get(resp.destactorid)
       ui.notifications.info(`${destActor.name} accepted ${resp.itemname}`)
     }